[Aymen] DocumentTerm modification and sortBasedIndexBuilder Creation

+++ IR/IR/DocumentTerm.h
@@ -1,5 +1,5 @@
 #pragma once
 typedef struct DocumentTerm {
-	unsigned long long documentIndex;
+	unsigned int documentIndex;
 	unsigned int ftd=1; // frequency of term in the current document
 } DocumentTerm;
